# openapi.json
{"openapi":"3.0.0","paths":{"/":{"get":{"operationId":"AppController_getHello","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"string"}}}}}}},"/users":{"get":{"operationId":"UserController_findAll","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"string"}}}}}},"put":{"operationId":"UserController_update","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserDto"}}}},"responses":{"200":{"description":""}}}},"/users/{id}":{"get":{"operationId":"UserController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"string"}}}}}},"delete":{"operationId":"UserController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"string"}}}}}}},"/authentication/login":{"post":{"operationId":"AuthenticationController_signIn","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"allOf":[{"$ref":"#/components/schemas/WebResponseDto"},{"properties":{"result":{"type":"object","properties":{"data":{"$ref":"#/components/schemas/AuthenticationTokenDto"}}}}}]}}}},"201":{"description":""},"400":{"description":"Returns \"Credentials not valid\" if authentication fails","content":{"application/json":{"schema":{"allOf":[{"$ref":"#/components/schemas/WebResponseDto"},{"properties":{"errors":{"type":"object","properties":{"code":{"type":"string"},"message":{"type":"string"}}}}}]}}}}}}},"/authentication/logout":{"post":{"operationId":"AuthenticationController_logout","parameters":[],"responses":{"201":{"description":""}}}},"/authentication/register":{"post":{"operationId":"AuthenticationController_signUp","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignUpDto"}}}},"responses":{"201":{"description":""}}}},"/authentication/generate-otp":{"get":{"operationId":"AuthenticationController_generateOneTimePasswordVerification","parameters":[],"responses":{"200":{"description":""},"201":{"description":"The OTP has been successfully sent","content":{"application/json":{"schema":{"$ref":"#/components/schemas/"}}}}}}},"/authentication/verify-otp":{"post":{"operationId":"AuthenticationController_verifyOneTimePasswordVerification","parameters":[],"responses":{"201":{"description":""}}}},"/authentication/reset-password":{"post":{"operationId":"AuthenticationController_resetPassword","parameters":[],"responses":{"201":{"description":""}}}},"/authentication/google":{"get":{"operationId":"AuthenticationController_googleAuth","parameters":[],"responses":{"200":{"description":""}}}},"/authentication/google-redirect":{"get":{"operationId":"AuthenticationController_googleAuthRedirect","parameters":[],"responses":{"200":{"description":""}}}},"/predict":{"post":{"operationId":"PredictController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreatePredictDto"}}}},"responses":{"201":{"description":""}}},"get":{"operationId":"PredictController_findAll","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"string"}}}}}}},"/predict/{id}":{"get":{"operationId":"PredictController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"string"}}}}}},"patch":{"operationId":"PredictController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdatePredictDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"string"}}}}}},"delete":{"operationId":"PredictController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"string"}}}}}}},"/questions":{"post":{"operationId":"QuestionController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateQuestionDto"}}}},"responses":{"201":{"description":""}}},"get":{"operationId":"QuestionController_findAll","parameters":[],"responses":{"200":{"description":""}}}},"/questions/{id}":{"get":{"operationId":"QuestionController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"string"}}}}}},"put":{"operationId":"QuestionController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateQuestionDto"}}}},"responses":{"200":{"description":""}}},"delete":{"operationId":"QuestionController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}}}}},"info":{"title":"SuaraNusa API Documentation","description":"SuaraNusa API Documentation - Bangkit 2024 Batch II","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"schemas":{"UpdateUserDto":{"type":"object","properties":{"name":{"type":"string"},"email":{"type":"string"},"password":{"type":"string"},"confirmPassword":{"type":"string"}},"required":["name","email","password","confirmPassword"]},"Paging":{"type":"object","properties":{"size":{"type":"number"},"totalPage":{"type":"number"},"currentPage":{"type":"number"}},"required":["size","totalPage","currentPage"]},"WebResponseDto":{"type":"object","properties":{"result":{"type":"object","nullable":true},"errors":{"type":"object","nullable":true},"paging":{"nullable":true,"allOf":[{"$ref":"#/components/schemas/Paging"}]}}},"AuthenticationTokenDto":{"type":"object","properties":{"accessToken":{"type":"string"},"refreshToken":{"type":"string"}},"required":["accessToken","refreshToken"]},"SignUpDto":{"type":"object","properties":{"name":{"type":"string"},"email":{"type":"string"},"password":{"type":"string"},"confirmPassword":{"type":"string"},"verificationQuestions":{"type":"array","items":{"type":"object"}}},"required":["name","email","password","confirmPassword","verificationQuestions"]},"CreatePredictDto":{"type":"object","properties":{}},"UpdatePredictDto":{"type":"object","properties":{}},"CreateQuestionDto":{"type":"object","properties":{"question":{"type":"string"}},"required":["question"]},"UpdateQuestionDto":{"type":"object","properties":{}}}}}